[r7] Add file with escape code zero, fix 24-bit overflow check
A file is added in which a rice partition is escaped to 0 bits,
so after the escape parameter and the 0, another rice partition
follows directly. This has not yet been implemented in any encoder
as far as I know.

Also, it was brought to my attention there was something wrong with
file 63: https://github.com/xiph/flac/issues/393 For the file to
decode correctly, the decoder relied on wrap-around of a 32-bit value.
The prediction + residual produced a value exceeding the range of
a 32-bit signed int, which was clipped to the correct value. This
was not intented, so the file is replaced with a correct one.

Also, change links from a to span in html playlist